refactor: decouple recorder
