+++ src-tauri/Cargo.toml
@@ -50,6 +50,7 @@ tauri-plugin-notification = "2"
 rand = "0.8.5"
 base64 = "0.21"
 mime_guess = "2.0"
+async-trait = "0.1.87"
 
 [features]
 # this feature is used for production builds or when `devPath` points to the filesystem
